+++ src/containers/item/Component.js
@@ -3,4 +3,5 @@
 export interface Component<T, B> {
   parent: T,
   builder(): B,
+  any(): boolean,
 }