+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 ## [Unreleased](https://github.com/eps1lon/poe-mods/compare/v1.4.0...HEAD)
+### Added
+- `Container#any` and `ItemComponent#any` helpers
 
 ## [1.4.0](https://github.com/eps1lon/poe-mods/compare/v1.3.5...v1.4.0) (2017-09-27)
 ### Breaking