+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
-## [Unreleased](https://github.com/eps1lon/poe-mods/compare/v1.4.0...HEAD)
+
+## [Unreleased](https://github.com/eps1lon/poe-mods/compare/v1.5.0...HEAD)
+
+## [1.5.0](https://github.com/eps1lon/poe-mods/compare/v1.4.0...v1.5.0) (2017-09-29)
 ### Breaking
 - Removed the `Stats` component from `Item` for Stats from Mods can directly be
   access via `#stats`. Things like ES or damage are now properties of